requirements-runtime.txt
typer>=0.9
pydantic>=2
pandas>=2.0
numpy>=1.26
jinja2>=3.1
nbconvert>=7.0
pyyaml>=6.0
loguru>=0.7
pyarrow>=15.0